+++ SICS/src/model/dimension/UnidimensionalModel.h
@@ -17,7 +17,7 @@ public:
 
 	// Methods
 	int getNumDimensions ();
-	vector<double> getDimVector();
+	vector<int> getDimVector();
 
 	// Getters and Setters
 	LatentTraitSet* getLatentTraitSet() const;
